ASCII WordArt to text Eval
Prompting GPT-3 with ASCII art has always been... weird. The model never refuses to generate art, but it always does it incorrectly. For example, by asking GPT-3.5. In this eval it was tried the other way around: GPT-3.5 is prompted with some ASCII-WordArt, and will show the words that it thinks the prompt represents.